Move ProcessCapture related utility functions
It makes more sense to keep them close to the class.
