+++ util/src/main/kotlin/Utils.kt
@@ -21,7 +21,6 @@ package com.here.ort.util
 
 import com.fasterxml.jackson.databind.JsonNode
 import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.databind.node.JsonNodeFactory
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory
 import com.fasterxml.jackson.module.kotlin.registerKotlinModule
 
@@ -83,56 +82,6 @@ object OkHttpClientHelper {
  * Return the directory to store user-specific configuration in.
  */
 fun getUserConfigDirectory() = File(System.getProperty("user.home"), ".ort")
-
-/**
- * Parse the standard output of a process as JSON.
- */
-fun parseJsonProcessOutput(workingDir: File, vararg command: String, jsonLines: Boolean = false): JsonNode {
-    val process = ProcessCapture(workingDir, *command).requireSuccess()
-
-    // Support parsing multiple lines with one JSON object per line by wrapping the whole output into a JSON array.
-    if (jsonLines) {
-        val array = JsonNodeFactory.instance.arrayNode()
-        process.stdoutFile.readLines().forEach { array.add(jsonMapper.readTree(it)) }
-        return array
-    }
-
-    return jsonMapper.readTree(process.stdout())
-}
-
-/**
- * Run a command to get its version.
- */
-fun getCommandVersion(command: String, versionArgument: String = "--version",
-                      semverType: Semver.SemverType = Semver.SemverType.LOOSE,
-                      transform: (String) -> String = { it }): Semver {
-    val version = ProcessCapture(command, versionArgument).requireSuccess()
-
-    var versionString = transform(version.stdout().trim())
-    if (versionString.isEmpty()) {
-        // Fall back to trying to read the version from stderr.
-        versionString = version.stderr().trim()
-    }
-
-    return Semver(versionString, semverType)
-}
-
-/**
- * Run a command to check it for specific version.
- */
-fun checkCommandVersion(command: String, expectedVersion: Semver, versionArgument: String = "--version",
-                        ignoreActualVersion: Boolean = false, transform: (String) -> String = { it }) {
-    val actualVersion = getCommandVersion(command, versionArgument, expectedVersion.type, transform)
-    if (actualVersion != expectedVersion) {
-        val messagePrefix = "Unsupported $command version $actualVersion, version $expectedVersion is "
-        if (ignoreActualVersion) {
-            println(messagePrefix + "expected.")
-            println("Still continuing because you chose to ignore the actual version.")
-        } else {
-            throw IOException(messagePrefix + "required.")
-        }
-    }
-}
 
 /**
  * Normalize a VCS URL by converting it to a common pattern. For example NPM defines some shortcuts for GitHub or GitLab
